scripts: set the interface hint when setting src/dst interface
The API that is used to send state to the controller has an 'interface'
parameter that is used when the config references an interface. The
controller will use this to order messages to the dataplane so that the
interface has always been created on the dataplane before the state that
refers to it is sent.  For the portmonitor commands for the src/dst
interface set the parameter when calling the controller instead of passing
in undef.

VRVDR-53049